SCOPE/requirements.txt
attr==0.3.2
attrs==21.2.0
numpy==1.21.2
OpenCC==1.1.2
pypinyin==0.43.0
python_Levenshtein==0.20.5
pytorch_lightning==1.5.1
regex==2021.10.8
tokenizers==0.10.3
torchmetrics==0.6.0
tqdm==4.62.3
transformers==4.12.5
